harman code

+++ programd77.cpp
@@ -5,6 +5,6 @@ int main(){
     int n;
     cout<<"enter the radius "<<endl;
     cin>>n;
-    int area = Pi*n*n;
+    float area = Pi*n*n;
     cout<<"area is "<<area<<endl;
 }
